Rename index.jsx to NameForm
- rename index.jsx to NameForm so that it is easier to find
